add args list & separate SmartPool
