+++ modules/runtime/src/misc/log/logger_queue.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "containers/concurrent_queue.h"
+
+namespace skr {
+namespace log {
+
+
+
+} } // namespace skr::log
